% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/All.R
\name{All}
\alias{All}
\title{All.}
\usage{
All(f, x, ..., na.rm = FALSE)
}
\arguments{
\item{f}{a predicate function.}

\item{x}{an object.}

\item{...}{further arguments passed to or from other methods.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
a logical value.
}
\description{
\code{All()} is a predicate functional that takes a predicate function
\code{f} and an object \code{x} and loops over each value \code{i} in
object \code{x}, evaluates \code{f(i)}, and returns TRUE if all values
\code{i} in object \code{x} evaluate as TRUE.
}
\examples{
# comment here
All(is.numeric, mtcars) # TRUE
All(is.character, mtcars) # FALSE
mtcars$am <- factor(mtcars$am)
All(is.numeric, mtcars) # FALSE
All(is.factor, mtcars) # FALSE

# Handles NAs and NULLs
All(is.numeric, list(NA, "3", NULL)) # FALSE
All(is.numeric, list(NA, 3, NULL)) # FALSE
All(is.numeric, list(NA, "3", NULL, 5)) # FALSE

# Use na.rm = TRUE to remove NULLS
All(Identity, list(NA, TRUE)) # NA
All(Identity, list(NA, TRUE), na.rm = TRUE) # TRUE
}

