% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LogicalOperators.R
\name{And}
\alias{And}
\title{Logical Operators}
\usage{
And()
}
\arguments{
\item{x,y}{need to edit}
}
\description{
These operators act on raw, logical, and number-like vectors
}
\examples{
# comment here
! x
x & y
x && y
x | y
x || y
xor(x, y)
isTRUE(x)
}

