% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Andify.R
\name{Andify}
\alias{Andify}
\title{Create new predicate functions}
\usage{
Andify(...)
}
\arguments{
\item{...}{n functions to apply in order from right to left}
}
\description{
Create new predicate functions
}
\examples{
# Examples
even_number_between_10_and_100 <- Andify(is.numeric, function(x) x > 10, function(x) x < 100, function(x) x \%\% 2 == 0)
even_number_between_10_and_100(8) # FALSE
even_number_between_10_and_100(9) # FALSE
even_number_between_10_and_100(10) # FALSE
even_number_between_10_and_100(11) # FALSE
even_number_between_10_and_100(12) # TRUE
even_number_between_10_and_100(49) # FALSE
even_number_between_10_and_100(50) # TRUE
even_number_between_10_and_100(100) # FALSE
even_number_between_10_and_100(101) # FALSE
even_number_between_10_and_100(102) # FALSE
# checker <- Andify(Compose(Not, is.numeric), Compose(Not, is.character))
}

