% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Apply}
\alias{Apply}
\title{Apply}
\usage{
Apply(f, x, margin, ...)
}
\arguments{
\item{f}{a function.}

\item{check}{a function.}

\item{init}{a valie.}
}
\value{
a list.
}
\description{
\code{Apply}
}
\examples{
# Iterate until the check condition is met
IterateUntil(function(n) { return(n + n) },
function(n) { return(n <= 1024) },
1)
}

