% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dollar.functionInfo.R
\name{$.functionInfo}
\alias{$.functionInfo}
\title{Using $ sign to access the functions within a functionInfo object}
\usage{
\method{$}{functionInfo}(x, name)
}
\arguments{
\item{name}{character, the element to extract}
}
\description{
Using $ sign to access the functions within a functionInfo object
}
\examples{
\dontrun{
library(Matrix)
library(ggplot2)
x = funCode(print)

x$default # access a function for ANY (default S3) objects
x$ggplot # access a function for S3 objects
x$diagonalMatrix # access a function for S4 objects
x$asdfasdfasdf # show a warning
}
}
