% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBiocReports.R
\name{checkBiocReports}
\alias{checkBiocReports}
\title{Check the package reports on Bioconductor}
\usage{
checkBiocReports(
  package = "RegEnrich",
  url0 = "http://bioconductor.org/checkResults/devel/bioc-LATEST/"
)
}
\arguments{
\item{package}{the package name}

\item{url0}{the url of the latest bioconductor package reports.}
}
\description{
Check the package reports on Bioconductor
}
\examples{
\dontrun{
checkBiocReports()
checkBiocReports("FlowSOM")
}
}
