% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Compact.R
\name{Compact}
\alias{Compact}
\title{Filter NA and NULL values out of a vector, list, or data.frame.}
\usage{
Compact(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Vector x but with all NULL and NA values filtered out.
}
\description{
\code{Compact()} takes a vector \code{x} and returns it with all NULL and
NA values filtered out.
}
\examples{
# Removes all null elements from a vector:
a <- list(NULL, 1, 5, NULL)
Compact(a)

b <- c(1, 2, 0, 4, NULL, 1, 3, NULL)
Compact(b)
}

