% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Compose.R
\name{Compose}
\alias{Compose}
\title{Compose}
\usage{
Compose(f, g)
}
\arguments{
\item{f}{a function.}

\item{g}{a function.}
}
\value{
a function.
}
\description{
\code{Compose()} takes two functions and composes them.
}
\examples{
# Compose a function using length and unique
lu <- Compose(length, unique)
lu(c(1:10, 5:15, 20:25))
}

