% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createProject}
\alias{createProject}
\title{Create a project folder, where it contains "0_RawData", "Rdata", "Rscript",
"figs", "csv", and "ppt" folders. And create two R script files
("00_01_functions.R" and "01_01_readData.R") in Rscript folder. If \code{.git = TRUE},
then a git repository will be initialized.}
\usage{
createProject(proj, path = "./", .git = FALSE, showWarnings = FALSE)
}
\arguments{
\item{proj}{project name}

\item{path}{the path where the project is created}

\item{.git}{logical, whether to create a git repository, default is FALSE}

\item{showWarnings}{logical, whether to show warnings when create the folders, default is FALSE}
}
\value{
a list of two elements, i.e., `dirs` and `files`. `dirs` contains all
the folders that have been created. `files` contains all the files that have
been created.
}
\description{
Create a project folder, where it contains "0_RawData", "Rdata", "Rscript",
"figs", "csv", and "ppt" folders. And create two R script files
("00_01_functions.R" and "01_01_readData.R") in Rscript folder. If \code{.git = TRUE},
then a git repository will be initialized.
}
\examples{
\dontrun{
# Create 'tempProj' project in the current working directory
createProject("tempProj")
dir.exists("tempProj")
unlink("tempProj",recursive = TRUE) # remove this folder
# Create 'tempProj2' project in the home directory
createProject("tempProj2", path = "~/")
dir.exists("~/tempProj2")
unlink("~/tempProj2",recursive = TRUE) # remove this folder
# Create 'tempProj3' project in the home directory and initialize it as a git repository.
# But you need to install git in the system in advance.
createProject("tempProj3", path = "~/", .git = TRUE)
dir.exists("~/tempProj3/.git")
unlink("~/tempProj3",recursive = TRUE) # remove this folder
}
}
