% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachPackage.R
\name{detachPackages}
\alias{detachPackages}
\title{detach packages}
\usage{
detachPackages(packages, prefix = "package:")
}
\arguments{
\item{packages}{package names}
}
\value{
removed package name(s)
}
\description{
detach packages
}
\examples{
\dontrun{
# detach one package
library(funcTools)
detachPackages("funcTools")

# detach multiple packages
library(funcTools)
library(ggplot2)
library(stringr)
detachPackages(c("funcTools", "ggplot2", "stringr"))
}
}
\seealso{
\code{\link{detachPackagesAll}}
}
