% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachPackage.R
\name{detachPackagesAll}
\alias{detachPackagesAll}
\title{detach all packages except basic packages}
\usage{
detachPackagesAll()
}
\value{
removed package name(s)
}
\description{
detach all packages except "base", "methods", "datasets",
"utils", "graphics", "stats", and "grDevices".
}
\examples{
\dontrun{
library(funcTools)
library(ggplot2)
library(stringr)
detachPackagesAll()
}
}
\seealso{
\code{\link{detachPackages}}
}
