% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge2Mat.R
\name{edge2Mat}
\alias{edge2Mat}
\title{Convert network edges to adjacency matrix}
\usage{
edge2Mat(edge, rownm, colnm, direction = FALSE, rmName = FALSE,
  restValue = 0)
}
\arguments{
\item{edge}{a data.frame of three columns, the first two columns
represent nodes in a network, the third column is the weight.}

\item{rownm}{row names in the final adjacency matrix. All node names
must present in either rownm or colnm.}

\item{colnm}{column names in the final adjacency matrix. All node names
must present in either rownm or colnm.}

\item{restValue}{the value to fill in the blank edges.}
}
\value{
An adjacency matrix.
}
\description{
Convert network edges to adjacency matrix
}
\examples{
\dontrun{
edge = data.frame(c("a", "b", "c"), c("a", "c", "d"), c(1,1,1))
edge2Mat(edge, edge[,1], edge[,2])

nodes = unique(c(as.character(edge[,1]), as.character(edge[,2])))
edge2Mat(edge, nodes, nodes)

edge = data.frame(c(1:3), c(3:5), c(1,1,1))
edge2Mat(edge,  edge[,1],  edge[,2] )
}
}
