% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Existy}
\alias{Existy}
\title{Existy}
\usage{
Existy(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
\code{Existy()}
}
\examples{
# Returns if a value exists or not:
Existy(4) # TRUE
Existy("foo) # TRUE
Existy(NULL) # FALSE
Existy(NA) # FALSE

# Works with lists
Exist(list(4, "foo", NULL, NA)) # TRUE
Existy(list(4, "foo")) # TRUE
Existy(list(NULL, NA)) # TRUE
Existy(list(NULL)) # TRUE
Existy(list(NA)) # FALSE

# Works with applying over lists
lapply(list(4, "foo", NULL, NA), Existy) # TRUE, TRUE, FALSE, FALSE
}

