% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Failwith}
\alias{Failwith}
\title{Failwith}
\usage{
Failwith(default = NULL, f, quiet = FALSE)
}
\arguments{
\item{f}{a predicate function.}
}
\value{
the negation of that function.
}
\description{
\code{Failwith()} turns a function that throws an error into a function that returns a default value when there’s an error.
The essence of failwith() is simple; it’s just a wrapper around try(), the function that captures errors and allows execution to continue.
}
\examples{
# Create a function, compact(), that removes all null elements from a list:
new_model <- lm(mtcars, formula = hp ~ wt)
getCoefficients <- Plucker("coefficients")
getCoefficients(new_model)
}

