% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{False}
\alias{False}
\title{False}
\usage{
False()
}
\arguments{
\item{f}{a predicate function.}
}
\value{
the negation of that function.
}
\description{
\code{False()}
}
\examples{
# Create a function, compact(), that removes all null elements from a list:
}

