% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq2datatable.R
\name{fastq2dataframe}
\alias{fastq2dataframe}
\title{Read fastq file and convert to a data frame}
\usage{
fastq2dataframe(fastqFile)
}
\arguments{
\item{fastqFile}{a fastq file name}
}
\description{
Read fastq file and convert to a data frame
}
\seealso{
\link{\code{fastq2dataframe_rev}}
}
