% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionInfo.R
\docType{class}
\name{fcnS-class}
\alias{fcnS-class}
\title{fcnS is an S4 class
fcnS includes a list of function and corresponding function S3/S4 method names info (MethodsFunction)}
\description{
fcnS is an S4 class
fcnS includes a list of function and corresponding function S3/S4 method names info (MethodsFunction)
}
