% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileNameFactory.R
\name{fileNameFactory}
\alias{fileNameFactory}
\title{A function factory to create a function that is used to create a file name with a certain extention.}
\usage{
fileNameFactory(path = "Rdata", extension = ".Rdata", createFolder = TRUE)
}
\arguments{
\item{path}{the path that is concatenated to a file name.}

\item{extension}{the file extension that is concatenated to a file name.}

\item{createFolder}{logical, whether to create the folder if the path doesn't exist.}
}
\value{
fileNameFactory returns a function that create a file name. That function
has two attributs, i.e., `path` and `extension`
}
\description{
A function factory to create a function that is used to create a file name with a certain extention.
}
\examples{
\dontrun{
Rdata = fileNameFactory(path = "Rdata/", ".Rdata")
Rdata
Rdata("asdf")

x = 10
pngs = fileNameFactory("figs/", ".png")
pngs
pngs("pngFig", x)

pdfs = fileNameFactory("pdfs/", ".png")
pdfs
pdfs("first/", "pngFile_", x)

# Once the function is created, the path/extension will not be changed by changing the attributes
attr(pngs, "path") = "figs_2"
pngs
pngs("asdf")
# You need to recreate the function with the new path
pngs2 = fileNameFactory("figs_2/", ".png")
pngs2("asdf")
}
}
