% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Finder}
\alias{Finder}
\title{Finder}
\usage{
Finder(x, f)
}
\arguments{
\item{x}{a vector.}
}
\value{
a two-argument function that returns a logical constant.
}
\description{
\code{Finder()} takes a vector and a two-argument function that returns a logical constant (either TRUE or FALSE),
and returns the that function reduced over that vector. It is assumed that:
* The two-argument function evaluates its two arguments and returns TRUE if the first argument is "better" than the second argument.
* The two-argument function knows how to "unwrap" its arguments.
}
\examples{
Some examples:
}

