% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionInfo.R
\name{funCode}
\alias{funCode}
\title{funCode function}
\usage{
funCode(f = character(), pattern = NULL, envir = topenv(parent.frame()))
}
\arguments{
\item{f}{a function name with or without quotation}

\item{pattern}{regular expression to match the methods}

\item{envir}{environment to search}
}
\value{
A functionInfo object, containing the function type,
        whether containing S3 or S4 methods, dirrect function source,
        S3 methods info, and S4 methods info.
}
\description{
funCode function is to obtain the source code of an R function
}
\examples{
# The source of print
funCode(print)
funCode("print")
# The source of funCode
funCode(funCode)
# The source of plot for data.frame
funCode("plot", "data.frame")
}
