% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionInfo.R
\docType{class}
\name{functionInfo-class}
\alias{functionInfo-class}
\alias{functionInfo}
\title{functionInfo is an S4 class and a function to generate functionInfo object
functionInfo function}
\arguments{
\item{typeof}{character, the result generated by typeof() function}

\item{S3S4}{logical, whether the results of .S3methods() and .S4methods() exist}

\item{fcn}{function, the result shown by print.function() directly}

\item{fS3}{fcnS object, the information of S3 methods}

\item{fS4}{fcnS object, the information of S4 mehtods}
}
\value{
a functionInfo object
}
\description{
functionInfo is an S4 class and a function to generate functionInfo object
functionInfo function
}
