% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\docType{package}
\name{functools}
\alias{functools}
\alias{functools-package}
\title{functools: Tools for Functional Programming in R}
\description{
functools: Tools for Functional Programming in R
}

