% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionInfo.R
\name{funUse}
\alias{funUse}
\title{apply either the dispatched S3 or S4 method of a function to an object}
\usage{
funUse(f, object, S3 = NULL, S4 = NULL, envir = topenv(parent.frame()), ...)
}
\arguments{
\item{f}{function name.}

\item{object}{the object that the function f is processing.}

\item{S3}{dispatched S3 method or "originalFunction".}

\item{S4}{dispatched S4 method.}

\item{envir}{the working environment.}
}
\description{
apply either the dispatched S3 or S4 method of a function to an object
}
\examples{
\dontrun{
funUse(print, "asdfasdfasd")
funUse(print, "asdfadsfasdf", S3 = "default")

fit <- lm(sepal ~ Petal.Length + Petal.Width + Species, data = datasets::iris)
funUse(print, fit)
funUse(print, fit, S3 = "lm")
print.lm(fit)

res = funUse(lm, object = sepal ~ Petal.Length + Petal.Width + Species, data = datasets::iris)
res
res = funUse(lm, object = sepal ~ Petal.Length + Petal.Width + Species,
             S3 = "originalFunction", data = datasets::iris)
res
lm(sepal ~ Petal.Length + Petal.Width + Species, data = datasets::iris)
}
}
