% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLoadedCFun.R
\name{getLoadedCFun}
\alias{getLoadedCFun}
\title{Get the information of loaded C functions/routines in a package or shared object (so or dll)}
\usage{
getLoadedCFun(package)
}
\arguments{
\item{package}{package name or shared object name}
}
\value{
DLLRegisteredRoutines or NULL if no C function is found
}
\description{
Get the information of loaded C functions/routines in a package or shared object (so or dll)
}
\examples{
\dontrun{
tmp = getLoadedCFun("base")
tmp
class(tmp)

library(Matrix)
getLoadedCFun("Matrix")

library(devtools)
getLoadedCFun("devtools")
}
}
