% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Value Matching}
\usage{
... \%notin\% NA
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors are not supported}
}
\value{
A vector of the same length as x.
}
\description{
%notin% is a more intuitive interface as a binary operator, which returns a logical vector indicating if there is a match or not for its left operand.
}
\examples{
# Some examples:
"bc" \%notin\% letters # TRUE
"b" \%notin\% letters # FALSE
"c" \%notin\% letters # FALSE
}

