% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub.R
\name{hpcInfo}
\alias{hpcInfo}
\title{Information of HPC computing nodes}
\usage{
hpcInfo()
}
\value{
a data frame, including 
· queuename, the queue name;
· qtype, the queue type - one of B(atch), I(nteractive), C(heckpointing), P(arallel);
· resv.used.tot., the number of used and available job slots;
· load_avg, the load average of the queue host or another load value;
· arch, the architecture of the queue host;
· stats, the state of the queue - one of u(nknown), a(larm), A(larm), 
  C(alendar suspended), s(uspended), S(ubordinate), d(isabled), D(isabled), 
  E(rror), c(configuration ambiguous), o(rphaned), P(reempted).
· etc.
}
\description{
Information of HPC computing nodes
}
\examples{
\dontrun{
hpcInfo()

View(hpcInfo())
}
}
