% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identicalSlots.R
\name{identicalSlots}
\alias{identicalSlots}
\title{Judge if the slots in two S4 objects are identical}
\usage{
identicalSlots(obj1, obj2, slots = NULL)
}
\arguments{
\item{obj1}{First S4 object}

\item{obj2}{Second S4 object}

\item{slots}{The slot names, the default is NULL, which means all of the slot
names in the first object.}
}
\value{
A list of two elements (FALSE and TRUE) if both exists. 
The slot names in the FALSE element are not identical between two objects. 
In contrast, the slot names in the TRUE element are identical between two 
objects.
}
\description{
Judge if the slots in two S4 objects are identical
}
\examples{
\dontrun{
library(Matrix)
i = c(1,3:8)
j = c(2,9,6:10)
x = 7 * (1:7)
(A = sparseMatrix(i, j, x = x))
identicalSlots(A, A)

y = x + 1
(B = sparseMatrix(i, j, x = y))
identicalSlots(A, B)
}
}
