% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Constants.R
\name{Identity}
\alias{Identity}
\title{Return the object passed to it.}
\usage{
Identity(.x)
}
\arguments{
\item{.x}{an object.}
}
\value{
the object.
}
\description{
\code{Identity()} returns itself.
}
\examples{
# Return itself:
Identity(5)
Identity(mean)
Identity(lm(data = mtcars, mpg ~ cyl))
}
\seealso{
Other constants: \code{\link{False}}; \code{\link{Na}};
  \code{\link{Null}}; \code{\link{True}}
}

