% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideOutList.R
\name{insideOutList}
\alias{insideOutList}
\title{Make a named list inside out.}
\usage{
insideOutList(namedList)
}
\arguments{
\item{namedList}{A named list.}
}
\value{
A list.
}
\description{
Make a named list inside out. Then the elements of the list
becomes the names of list, and the names of list becomes its element.
}
\examples{
{
\dontrun{
a = list(a = c("a1", "a2", "a3"),
         b = c("a2", "b1", "b3"),
         d = c("d3", "b1", "a2"))
insideOutList(a)
}
}
}
