% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Iterate-Until.R
\name{Iterate_Until}
\alias{Iterate_Until}
\title{Iterate until conpletion.}
\usage{
Iterate_Until(f, check, init)
}
\arguments{
\item{f}{a function.}

\item{check}{a function.}

\item{init}{a valie.}
}
\value{
a list.
}
\description{
\code{Iterate_Until}
}
\examples{
# Iterate until the check condition is met
Iterate_Until(function(n) { return(n + n) },
function(n) { return(n <= 1024) },
1)
}

