% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lapply.R
\name{Lapply}
\alias{Lapply}
\title{Wrapper for lapply function.}
\usage{
Lapply(f, x, ...)
}
\arguments{
\item{f}{A function.}

\item{x}{A value.}

\item{...}{Optional arguments to f.}
}
\description{
Wrapper for \code{\link[base]{lapply}}.
}
\seealso{
\code{\link[base]{lapply}} for code and documentation.

Other functionals: \code{\link{Apply}};
  \code{\link{Mapply}}
}

