% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAsList.R
\name{loadAsList}
\alias{loadAsList}
\title{Load data and assign it/them into a list}
\usage{
loadAsList(file)
}
\arguments{
\item{file}{the .Rdata file}
}
\description{
Load data and assign it/them into a list
}
\examples{
\dontrun{
x = "a"
y = 1
save(x,y, file = "~/Downloads/tmp_xy.Rdata")
rm(list = c("x", "y"))
loadAsList("~/Downloads/tmp_xy.Rdata")
}
}
