% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsAccessibleVariable.R
\name{lsAccessibleVariable}
\alias{lsAccessibleVariable}
\title{list accessible variables of a function}
\usage{
lsAccessibleVariable(fun, printValue = FALSE, variable2print = NULL)
}
\arguments{
\item{fun}{function object.}

\item{printValue}{logical, if TRUE, print the value of each variable that
the \code{fun} can access. The default is FALSE.}

\item{variable2print}{character or vector of charactor or NULL. If it's NULL and
\code{printValue} is TRUE, the value of all variables will be printed. If this
is a (vector of) character and \code{printValue} is TRUE, then the value(s) of
variable(s) indicated by this (vector of) character will be printed. When there
is no such variable indicated by variable2print, a warning will be raised.}
}
\value{
the name(s) of variable(s) which can be accessed by \code{'fun'} function.
}
\description{
List the variables that a function can access
}
\examples{
\dontrun{
a = 1
f = function(b) {
  ab = 2
  innerF = function(c) ab * b * c
  return(innerF)
}
funObj = f(2)

lsAccessibleVariable(f) # global environment.
lsAccessibleVariable(funObj) # environment within function f.
lsAccessibleVariable(funObj, T) # all variables in environment within function f.
lsAccessibleVariable(funObj, T, "ab") # variables 'ab' in environment within function f.
}
}
