% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsClass.R
\name{lsClass}
\alias{lsClass}
\title{List all classes in packages}
\usage{
lsClass(packages = NULL, pattern = NULL, asList = FALSE)
}
\arguments{
\item{packages}{character or NULL, in which packages to list all classes.
Default is NULL, which means in all attached packages and the global 
environment.}

\item{pattern}{character, only the classes with the patter are listed. 
Default is NULL, meaning return all classes if exist.}

\item{asList}{logical, whether return a list object of all classes in 
each package, or by default (FALSE), return a data frame of three columns,
namely, index, package, and class.}
}
\description{
List all classes in packages
}
\examples{
\dontrun{
lsClass()
lsClass(pattern = "*functionInfo")
lsClass(asList = TRUE)

lsClass("funcTools")
lsClass("package:funcTools")
lsClass("methods")

lsClass("DESeq2")
lsClass(c("funcTools", "DESeq2"))
lsClass(packages = c("funcTools", "DESeq2", "asdf"))
lsClass(packages = c("funcTools", "DESeq2", "asdf", "asdf32"))
}
}
