% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnimation.R
\name{makeAnimation}
\alias{makeAnimation}
\title{Use the image files to create a gif animation}
\usage{
makeAnimation(
  imgs,
  pattern = NULL,
  path = NULL,
  fps = 2,
  outputgif = "output.gif"
)
}
\arguments{
\item{imgs}{image file full names}

\item{pattern}{if imgs is missing, then the files with this pattern in the 
names will be included from path.}

\item{path}{image file path}

\item{fps}{number of frames per second}

\item{outputgif}{output gif file names}
}
\description{
Use the image files to create a gif animation
}
