% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFilenames.R
\name{makeFilenames}
\alias{makeFilenames}
\title{[!+] Convert string into Windows compatible filename}
\usage{
makeFilenames(s, replacement = "_", allow.space = TRUE)
}
\arguments{
\item{s}{A string.}

\item{replacement}{A symbol to replce unallowed symbols with.
\code{replacement} must be an allowed symbol.}

\item{allow.space}{Logical. If \code{FALSE}, space will be
repalaced with \code{replacement}. Default is \code{TRUE}.}
}
\value{
A string without special symbols.
}
\description{
This is a function (make.filenames) from https://github.com/GegznaV/spMisc 
repository.
This function replaces special symbols (\code{\"  / : * ?  < > |}) in
a string \code{s} with \code{replacement} (by default it is underscore
 \code{_}). After replacement the  string can be used as a file name
 in Windows.
}
\examples{
 s <- '\\\\ / : * ? " < > |'
 makeFilenames(s)
 #> "_ _ _ _ _ _ _ _ _"

 makeFilenames(s, allow.space = FALSE)
 #> "_________________"


 s2 <- "/Hello?"
 makeFilenames(s2)
 #> "_Hello_"

 makeFilenames(s2, replacement = "-")
 #> "-Hello-"

}
