% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUniqueNames.R
\name{makeUniqueNames}
\alias{makeUniqueNames}
\title{Make Unique Names\cr}
\usage{
makeUniqueNames(x, sep = "_")
}
\arguments{
\item{x}{character vector for names (usually with duplicates).}

\item{sep}{a character to seperate the original character and added
an integer.}
}
\value{
A character vector without duplicates.
}
\description{
make an character vector suitable for names by adding
an integer to the end of the characters.
}
