% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Map.R
\name{Map}
\alias{Map}
\title{Wrapper for Map function.}
\usage{
Map(..., .f)
}
\arguments{
\item{...}{Arguments to vectorize over (vectors or lists of strictly positive length, or all of zero length).}

\item{.f}{A function to be applied.}
}
\description{
Wrapper for \code{\link[base]{Map}}.
}
\seealso{
\code{\link[base]{Map}} for code and documentation.

Other functionals: \code{\link{Apply}};
  \code{\link{Lapply}}; \code{\link{Mapply}};
  \code{\link{Sapply}}; \code{\link{Tapply}};
  \code{\link{Vapply}}
}

