% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Mapply.R
\name{Mapply}
\alias{Mapply}
\title{Wrapper for mapply function.}
\usage{
Mapply(f, x, ..., more_args = NULL, simplify = TRUE, use_names = TRUE)
}
\arguments{
\item{f}{A function.}

\item{...}{Arguments to vectorize over (vectors or lists of strictly positive length, or all of zero length).}

\item{more_args}{A list of other arguments to FUN.}

\item{simplify}{Logical or character string; attempt to reduce the result to a vector, matrix or higher dimensional array; see the simplify argument of \code{\link[base]{sapply}}.}

\item{use_names}{Logical; use names if the first ... argument has names, or if it is a character vector, use that character vector as the names.}
}
\description{
Wrapper for \code{\link[base]{mapply}}.
}
\seealso{
\code{\link[base]{mapply}} for code and documentation.

Other functionals: \code{\link{Apply}};
  \code{\link{Lapply}}
}

