% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge2Mat.R
\name{mat2Edge}
\alias{mat2Edge}
\title{Adjacency matrix to a data.frame of edges.}
\usage{
mat2Edge(mat, mode = c("directed", "undirected", "upper", "lower"),
  diag = FALSE)
}
\arguments{
\item{mat}{adjacency matrix.}

\item{mode}{Character, to specify the class of graph and which part of
the matrix will be used. Possible values are: "directed" (default),
"undirected", "upper", "lower".}

\item{diag}{logic, whether to include the diagonal of the matrix.}
}
\value{
a data.frame of edge information. The first column is from node,
the second column is to node, and the third is weight.
}
\description{
Adjacency matrix to a data.frame of edges.
}
\examples{
{
\dontrun{
mat = matrix(rnorm(4*4), nrow = 4,
             dimnames = list(letters[1:4], LETTERS[1:4]))
mat2Edge(mat, mode = "undirected", diag = TRUE)
mat2Edge(mat, mode = "undirected", diag = FALSE)
mat2Edge(mat, mode = "directed", diag = TRUE)
mat2Edge(mat, mode = "upper", diag = TRUE)
mat2Edge(mat, mode = "upper", diag = FALSE)
}
}
}
\seealso{
edge2Mat2
}
