% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixRep.R
\name{matrixRep}
\alias{matrixRep}
\title{Repeat matrix x by m rows and n columns}
\usage{
matrixRep(x, m = 1, n = 1)
}
\arguments{
\item{x}{a matrix or a numeric vector.}

\item{m}{the desired number of rows.}

\item{n}{the desired number of columns.}
}
\value{
Return a matrix repeated m (rows) * n (columns) times
}
\description{
Repeat matrix x by m rows and n columns
}
\examples{
{
\dontrun{
matrixRep(rnorm(3), m = 2, n = 3)
matrixRep(matrix(rnorm(6), ncol = 2), m = 3, n = 4)

matrixRep(t(c(T, F)), m = 4, n = 3)
matrixRep(t(letters[1:5]), m = 4, n = 1)
}
}
}
\seealso{
\code{\link{matrixNew}}
}
