% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Memoise.R
\name{Memoise}
\alias{Memoise}
\title{Memoise a function.}
\usage{
Memoise(.f)
}
\arguments{
\item{.f}{Function of which to create a memoised copy.}
}
\value{
A memoised copy of the original function.
}
\description{
Wrapper for \code{\link[memoise]{memoise}}.
}
\seealso{
\code{\link[memoise]{memoise}} for code and documentation.

Other function operators: \code{\link{Compose}},
  \code{\link{\%O\%}}; \code{\link{Fail_With}};
  \code{\link{Partial}}; \code{\link{Reduce_Right}}
}

