% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Negate.R
\name{Negate}
\alias{Negate}
\title{Wrapper for Negate function.}
\usage{
Negate(.f)
}
\arguments{
\item{.f}{A function to be applied.}
}
\description{
Wrapper for \code{\link[base]{Negate}}.
}
\seealso{
\code{\link[base]{Negate}} for code and documentation.

Other functionals: \code{\link{Apply}};
  \code{\link{Lapply}}; \code{\link{Mapply}};
  \code{\link{Sapply}}; \code{\link{Tapply}};
  \code{\link{Vapply}}
}

