% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Constants.R
\name{Null}
\alias{Null}
\title{Return the value NULL.}
\usage{
Null()
}
\value{
NULL.
}
\description{
\code{Null()} is a function that returns NULL.
}
\examples{
# Null() returns NULL:
Null()
}
\seealso{
Other constants: \code{\link{False}};
  \code{\link{Identity}}; \code{\link{Na}};
  \code{\link{True}}
}

