% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Or.R
\name{Or}
\alias{Or}
\title{Or}
\usage{
Or(f, g)
}
\arguments{
\item{f}{a predicate function.}

\item{g}{a predicate function.}
}
\value{
a function.
}
\description{
\code{Or()}
}
\examples{
# comment here
}

