% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Orify.R
\name{Orify}
\alias{Orify}
\title{Orify}
\usage{
Orify(f, g)
}
\arguments{
\item{f}{a predicate function.}

\item{g}{a predicate function.}
}
\value{
a function.
}
\description{
\code{Orify()}
}
\examples{
# comment here
}

