% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parfor.R
\name{parfor}
\alias{parfor}
\title{parallel computing with a progress bar}
\usage{
parfor(x, FUN, ..., .core = 2, .printConsole = TRUE, .progressBar = TRUE)
}
\arguments{
\item{x}{a vector/list that function FUN needs}

\item{FUN}{function}

\item{...}{other parameters that function FUN needs}

\item{.core}{number of cores}

\item{.printConsole}{logical, if messages from FUN is printed}

\item{.progressBar}{logical, if a progressBar is added}
}
\description{
parallel computing with a progress bar
}
\examples{
\dontrun{
parfor(1:10, c, 1)
parfor(1:10, Sys.sleep, .core = 10)
y = parfor(1:10, print, .printConsole = TRUE, .progressBar = FALSE)
parfor(1:10, print, .printConsole = TRUE, .progressBar = TRUE)
}

}
