% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Partial.R
\name{Partial}
\alias{Partial}
\title{Partial apply a function, filling in some arguments.}
\usage{
Partial(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[pryr]{partial}}.}
}
\description{
Wrapper for \code{\link[pryr]{partial}}.
}
\seealso{
\code{\link[pryr]{partial}} for code and documentation.

Other function operators: \code{\link{Compose}},
  \code{\link{\%O\%}}; \code{\link{Fail_With}};
  \code{\link{Memoise}}
}

