% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste2strings.R
\name{\%,\%}
\alias{\%,\%}
\alias{\%, \%}
\alias{\%/\%}
\title{Paste two strings without a space}
\usage{
x \%,\% y

x \%, \% y

x \%/\% y
}
\arguments{
\item{x}{character}

\item{y}{character}
}
\value{
Paste two strings without a space

Paste two strings with a space

Paste two strings with a slash (/)
}
\description{
Paste two strings without a space

Paste two strings with a space

Paste two strings with a space
}
\examples{
{
"a" \%,\% "b"
"a" \%, \% "b"
"folder" \%/\% "file"
}
}
