% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSave.R
\name{plotSave}
\alias{plotSave}
\title{Save a plot}
\usage{
plotSave(filename, plotCMD, device = NULL, path = NULL, scale = 1,
  width = NA, height = NA, units = c("in", "cm", "mm"), dpi = 300,
  limitsize = TRUE, ...)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plotCMD}{Plot based on this plot command, cannot use a return
value of a plot command. for example define h = hist(rnorm(100)) and then
use plotCMD = h.}

\item{device}{Device to use. Can be either be a device function
(e.g. [png()]), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{path}{Path to save plot to (combined with filename).}

\item{scale}{Multiplicative scaling factor.}

\item{width, height, units}{Plot size in `units` ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{limitsize}{When `TRUE` (the default), `plotSave` will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}

\item{...}{Other arguments passed on to graphics `device`.}

\item{returnPlot}{When `FALSE` (the default), plotSave will not return
the value that plotCMD might return.}
}
\description{
`plotSave()` is a convenient function for saving a plot. It saves
the plot that the plotCMD is to make, using the size of the current
graphics device by default. It also guesses the type of graphics device
from the extension. This function is very convinent when saving a plot
as a pdf or svg file and the letters in the plot need to be able to copied.
}
\seealso{
Similar usage can refer to \code{\link{ggplot}()}.
}
