% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plucker.R
\name{Plucker}
\alias{Plucker}
\title{Create a function that plucks a field from S3 and S4 objects.}
\usage{
Plucker(field)
}
\arguments{
\item{field}{a character vector.}
}
\value{
a function that plucks fields from S3 and S4 object slots.
}
\description{
\code{Plucker} is a convenience function that when passed a character
vector as an argument, will return a function that "plucks" objects
from S3 and S4 object slots!
}
\examples{
# Pluck a value from a list or S3 object:
new_model <- lm(mtcars, formula = hp ~ wt)
getCoefficients <- Plucker("coefficients")
getCoefficients(new_model)
}
\seealso{
Other closures: \code{\link{Always}}
}

