% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub.R
\name{print.HpcInfo}
\alias{print.HpcInfo}
\title{print HpcInfo object}
\usage{
\method{print}{HpcInfo}(x, ...)
}
\arguments{
\item{x}{HpcInfo object.}

\item{...}{other parameters in print.data.frame function.}
}
\value{
print.HpcInfo returns an invisible data frame, which shows the 
available jobs, slots and memory on HPC nodes.
}
\description{
print HpcInfo object
}
