% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub.R
\name{print.statRes}
\alias{print.statRes}
\title{Print statRes object}
\usage{
\method{print}{statRes}(statRes)
}
\arguments{
\item{statRes}{the result of qstat/qstatAll/qstatGroupAll functions.}
}
\description{
Print statRes object
}
\examples{
\dontrun{
tmp = qstatAll()
print(tmp)
}
}
