% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub_slurm.R
\name{print.statRes2}
\alias{print.statRes2}
\title{Print statRes2 object}
\usage{
\method{print}{statRes2}(statRes)
}
\arguments{
\item{statRes}{the result of sacct/sacctAll/sacctGroupAll functions.}
}
\description{
Print statRes2 object
}
\examples{
\dontrun{
tmp = qstatAll2()
print(tmp)
}
}
