% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoLog.R
\name{pseudoLog}
\alias{pseudoLog}
\title{The Signed Pseudo Logarithm}
\usage{
pseudoLog(x, base = 10, sigma = 1, inverse = FALSE)
}
\arguments{
\item{x}{numeric value or vector that need to transform.}

\item{base}{base of logarithm.}

\item{inverse}{whether the inverse function of pseudoLog is applied.}

\item{Scaling}{factor for the linear part of pseudo-log transformation.}
}
\description{
The domain of logarithm function is >0. However, sometimes it is required to
transform numbers using sort of "signed logarithm". The pseudoLog function
can achieve this goal.
}
\examples{
\dontrun{
x = c(-10^(4:1), 0, 10^(1:4))
y = pseudoLog(x)
print(y)
z = pseudoLog(x, inverse = TRUE)
}

}
