% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub.R
\name{qdelAll}
\alias{qdelAll}
\title{Delete (qdel) of job by pattern}
\usage{
qdelAll(pattern = "*", col = 1)
}
\arguments{
\item{pattern}{The pattern to match.}

\item{col}{The column index to match pattern.}
}
\value{
Deleted job IDs.
}
\description{
Delete (qdel) of job by pattern
}
\seealso{
\code{\link{qdel}}
}
