% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub.R
\name{qstat}
\alias{qstat}
\title{Status (qstat) of my running jobs.}
\usage{
qstat(stat = c("all", "run", "wait"))
}
\arguments{
\item{stat}{the job status, including "all" (the default),
"run", and "wait".}
}
\description{
Status (qstat) of my running jobs.
}
\seealso{
\code{\link{qstatAll}}, \code{\link{qstatGroupAll}}
}
