% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub_slurm.R
\name{qstat2}
\alias{qstat2}
\title{checking your own job status on HPC for slurm system}
\usage{
qstat2(
  stat = c("run", "all", "wait", "COMPLETED", "TIMEOUT", "FAILED", "OUT_OF_MEMORY",
    "CANCELLED")
)
}
\arguments{
\item{stat}{status, one of "run", "all", "wait", 
"COMPLETED", "TIMEOUT", "FAILED", "OUT_OF_MEMORY", and "CANCELLED".}
}
\description{
checking your own job status on HPC for slurm system
}
\examples{
\dontrun{
qstat2(stat = "run")
qstat2("all")
}
}
