% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub.R
\name{qstatGroupAll}
\alias{qstatGroupAll}
\title{Status (qstat) of running jobs of group members.}
\usage{
qstatGroupAll(stat = c("run", "all", "wait"),
  groupMembers = system("ls /hpc/dla_lti", intern = TRUE))
}
\arguments{
\item{stat}{the job status, including "run" (the default),
"all", and "wait".}

\item{groupMembers}{the group member name.}
}
\description{
Status (qstat) of running jobs of group members.
}
\seealso{
\code{\link{qstat}}, \code{\link{qstatAll}}
}
