% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub_slurm.R
\name{qstatGroupAll2}
\alias{qstatGroupAll2}
\title{Status (sacct, equivalent to qstat) of running jobs of group members.}
\usage{
qstatGroupAll2(
  stat = c("run", "all", "wait", "COMPLETED", "TIMEOUT", "FAILED", "OUT_OF_MEMORY",
    "CANCELLED"),
  groupMembers = dir(paste0("/hpc/", strSplit(system("groups", intern = TRUE), " ")[1,
    1]))
)
}
\arguments{
\item{stat}{the job status, including "run" (the default),
"all", "wait", "COMPLETED", "TIMEOUT", "FAILED", "OUT_OF_MEMORY", 
and "CANCELLED".}

\item{groupMembers}{the group member names. The default is the group members 
of the user who run this code.}
}
\description{
Status (sacct, equivalent to qstat) of running jobs of group members.
}
\seealso{
\code{\link{qstat2}}, \code{\link{qstatAll2}}
}
