% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub.R
\name{qstatProcess}
\alias{qstatProcess}
\title{process the results form qstat command}
\usage{
qstatProcess(statRes)
}
\arguments{
\item{statRes}{the results from `system("qstat -r", TRUE)`.}
}
\value{
a data.frame of tidy results
}
\description{
process the results form qstat command
}
\examples{
\dontrun{
x = system("qstat -r", T)
qstatProcess(x)
}
}
