% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub_slurm.R
\name{qstatProcess2}
\alias{qstatProcess2}
\title{Status (sacct) of SLURM. jobs.}
\usage{
qstatProcess2(statRes)
}
\arguments{
\item{stat}{the job status, including "running", "completed", "failed",
"timeout", "resizing", "deadline", "node_fail".}
}
\description{
Status (sacct) of SLURM. jobs.
}
\seealso{
\code{\link{qstat}}
}
