% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub.R
\name{qstatSummary}
\alias{qstatSummary}
\title{Summarize qstat results}
\usage{
qstatSummary(statRes)
}
\arguments{
\item{statRes}{the result of qstat/qstatAll/qstatGroupAll functions.}
}
\description{
Summarize qstat results
}
\examples{
\dontrun{
tmp = qstatAll()
qstatSummary(tmp)
}
}
