% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_switch.R
\name{R_currentFramework}
\alias{R_currentFramework}
\title{The current version of R framework}
\usage{
R_currentFramework(frameworkPath = "/Library/Frameworks/R.framework/")
}
\arguments{
\item{frameworkPath}{the R.framework path, the default is 
"/Library/Frameworks/R.framework/" on Mac.}
}
\description{
The current version of R framework
}
\examples{
\dontrun{
R_currentFramework()
}
}
