% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchMacR.R
\name{R_frameworks}
\alias{R_frameworks}
\title{list all R frameworks}
\usage{
R_frameworks(frameworkPath = "/Library/Frameworks/R.framework/")
}
\arguments{
\item{frameworkPath}{the R.framework path, the default is 
"/Library/Frameworks/R.framework/" on Mac.}
}
\description{
list all R frameworks
}
\examples{
\dontrun{
R_frameworks()
}
}
