% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchMacR.R
\name{R_switch}
\alias{R_switch}
\title{Switch R framework version on Mac}
\usage{
R_switch(
  version = R_frameworks(),
  reset = TRUE,
  frameworkPath = "/Library/Frameworks/R.framework/",
  RStudioPath = "/Applications/RStudio.app/Contents/MacOS/RStudio"
)
}
\arguments{
\item{reset}{if TRUE (default), the framework is changed back to the
original framework version.}

\item{frameworkPath}{the R.framework path, the default is 
"/Library/Frameworks/R.framework/" on Mac.}

\item{RStudioPath}{RStudio path, the default is 
"/Applications/RStudio.app/Contents/MacOS/RStudio".}

\item{Version}{R framework version to switch to}
}
\description{
Switch R framework version on Mac
}
\examples{
\dontrun{
# List all frameworks
fw = R_frameworks()
fw

# Current framework
cfw = R_currentFramework()
cfw

# Switch
R_switch(fw[1], reset = FALSE)
R_switch(fw[2])

# Reset the initial framework
R_switch(cfw, reset = FALSE)

}
}
