% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_switch.R
\name{R_version}
\alias{R_version}
\title{Current R version}
\usage{
R_version(full = FALSE)
}
\arguments{
\item{full}{logical, if TRUE, the date will be shown. If the R is a
development version, then "devel" will be added ahead of the version
number.}
}
\description{
Current R version
}
\examples{
\dontrun{
R_version()
R_version(TRUE)
}

}
