% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Reduce-Right.R
\name{Reduce_Right}
\alias{Reduce_Right}
\title{Reduce Right}
\usage{
Reduce_Right(f, x, ...)
}
\arguments{
\item{f}{A predicate function.}

\item{x}{A vector.}

\item{...}{Further arguments passed to the predicate function.}
}
\value{
To be updated.
}
\description{
Reduce Right
}

