% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Reject}
\alias{Reject}
\title{Reject}
\usage{
Reject(f, x)
}
\arguments{
\item{x}{a vector.}
}
\description{
\code{Reject()} takes a vector x and returns it with all NULL values filtered out.
}
\examples{
Some examples:
}

