% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub_slurm.R
\name{sacct}
\alias{sacct}
\title{sacct: checking your own job status on HPC for slurm system}
\usage{
sacct(
  stat = c("run", "all", "wait", "COMPLETED", "TIMEOUT", "FAILED", "OUT_OF_MEMORY",
    "CANCELLED")
)
}
\arguments{
\item{stat}{status, one of "run", "all", "wait", 
"COMPLETED", "TIMEOUT", "FAILED", "OUT_OF_MEMORY", and "CANCELLED".}
}
\description{
sacct: checking your own job status on HPC for slurm system
}
\examples{
\dontrun{
sacct(stat = "run")
sacct("all")
}
}
