% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub_slurm.R
\name{sacctAll}
\alias{sacctAll}
\title{sacctAll: checking all users' job status on HPC for slurm system}
\usage{
sacctAll(
  stat = c("run", "all", "wait", "COMPLETED", "TIMEOUT", "FAILED", "OUT_OF_MEMORY",
    "CANCELLED")
)
}
\arguments{
\item{stat}{status, one of "run", "all", "wait", 
"COMPLETED", "TIMEOUT", "FAILED", "OUT_OF_MEMORY", and "CANCELLED".}
}
\description{
sacctAll: checking all users' job status on HPC for slurm system
}
\examples{
\dontrun{
sacctAll("all")
sacctAll("run")
sacctAll("TIMEOUT")
}
}
