% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleReads.R
\name{sampleReads}
\alias{sampleReads}
\title{sampleReads}
\usage{
sampleReads(reads, size = 10^5, rep = 10, prob = "dunif", sd = 0,
  addNoise = FALSE)
}
\arguments{
\item{reads}{Vector of gene names (factor or character) or a data frame
(or matrix) of gene names and corresponding frequencies.}

\item{size}{Integer, the number of reads of downsampling.}

\item{rep}{Integer, the number of sampling to mimic the number of sigle
cells.}

\item{prob}{Either "dunif" or the numeric vector of gene reads representing
the prio probability to do sampling.}

\item{sd}{***}

\item{addNoise}{Logical. If TRUE, add extra noise to the sampled data}
}
\value{
A data frame, the row names of which are gene names, and each column
of which is the frequencies of genes in a single cell.
}
\description{
Provide gene reads of bulk sequencing data, do downsampling
to get a new set of gene reads to mimic the single cell sequencing data.
}
