% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Sapply.R
\name{Sapply}
\alias{Sapply}
\title{Wrapper for sapply function.}
\usage{
Sapply(f, x, ..., simplify = TRUE, use_names = TRUE)
}
\arguments{
\item{f}{A function to be applied.}

\item{x}{A vector.}

\item{...}{Optional arguments to f.}

\item{simplify}{Logical or character string; should the result be simplified to a vector, matrix or higher dimensional array if possible? For sapply it must be named and not abbreviated. The default value, TRUE, returns a vector or matrix if appropriate, whereas if simplify = "array" the result may be an array of 'rank' (=length(dim(.))) one higher than the result of FUN(X[[i]]).}

\item{use_names}{Logical; if TRUE and if X is character, use X as names for the result unless it had names already. Since this argument follows ... its name cannot be abbreviated.}
}
\description{
Wrapper for \code{\link[base]{sapply}}.
}
\seealso{
\code{\link[base]{sapply}} for code and documentation.

Other functionals: \code{\link{Apply}};
  \code{\link{Lapply}}; \code{\link{Mapply}};
  \code{\link{Tapply}}; \code{\link{Vapply}}
}

