% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_xy_ticks}
\alias{scale_xy_ticks}
\title{Change the x and y axis ticks of ggplot}
\usage{
scale_xy_ticks(gplot, xBreaks = NULL, yBreaks = NULL)
}
\arguments{
\item{gplot}{a ggplot object.}

\item{xBreaks}{the x axis ticks.}

\item{yBreaks}{the y axis ticks.}
}
\value{
a ggplot object with changed x and y axis ticks.
}
\description{
Change the x and y axis ticks of ggplot
}
\examples{
\dontrun{
dat = data.frame(x = rnorm(100), y = rnorm(100))
g = ggplot(dat, aes(x,y)) +
  geom_point()
scale_xy_ticks(g, pretty(dat$x, n = 10), pretty(dat$y, n = 10))
}
}
