% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub_slurm.R
\name{scancel}
\alias{scancel}
\title{scancel: Delete (scancel) of SLURM jobs.}
\usage{
scancel(id)
}
\arguments{
\item{id}{The job ID.}
}
\description{
scancel: Delete (scancel) of SLURM jobs.
}
\seealso{
\code{\link{scancelAll}}
}
