% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd2.R
\name{setwd2}
\alias{setwd2}
\title{setwd2}
\usage{
setwd2(dir, preDir = NULL, local = "Ubuntu", remote = "CentOS")
}
\arguments{
\item{dir}{A character string: path started from the root.}

\item{preDir}{A character string: if the remote path is accessed locally,
what path needs to be prefixed? If preDir = NULL, only dir is considered.}

\item{local}{A character string: local platform name, which can be
obtained by running \code{\link{sessionInfo}}()$running. If local = NULL,
then preDir will not be considered. Letter case is ignored.}

\item{remote}{A character string: remote platform name,
if NULL then preDir will not be considered. Letter case is ignored.}
}
\value{
The present working directory before setting new one.
}
\description{
setwd2 is to set working directory in case the folder
is on remote computer (e.g. HPC) but accessed locally.
}
\examples{
# set working directory as the local home
setwd2("~")

# set working directory as remote home mounting path if remote is matched
\dontrun{setwd2(dir = "/Users/homename", preDir = "[mounting path]",
    local = "ubuntu", remote = "centos")}
}
\seealso{
\code{\link{setwd}}
}
