% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd2thisFile.R
\name{setwd2thisFile}
\alias{setwd2thisFile}
\title{Set working directory to where the file that put this
function located.}
\usage{
setwd2thisFile()
}
\value{
the working directory before the change.
}
\description{
Set working directory to where the file that put this
function located.
}
\examples{
{
\dontrun{
# Put the following codes in one file and save.
getwd()
setwd2thisFile()
getwd()
}
}
}
