% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionInfo.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,functionInfo-method}
\title{show,functionInfo-method}
\usage{
\S4method{show}{functionInfo}(object)
}
\arguments{
\item{object}{an object of functionInfo class}
}
