% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slotClass.R
\name{slotClass}
\alias{slotClass}
\title{Obtain class names of slots in an S4 object}
\usage{
slotClass(object, slot_names = NULL)
}
\arguments{
\item{object}{an S4 object}

\item{slot_names}{character, the class of which slot to obtain. Default is NULL, 
meaning obtain the class name of all slots that object has.}
}
\value{
a list of class names for either all slots (if slot_names = NULL) or for
the slots that `slot_names` specifys.
}
\description{
get class names of slots in an S4 object, which is similar 
to \link{\code{getSlots}} function, but \code{getSlots} get class names of 
slots based on a class name not an object instance.
}
\examples{
\dontrun{
## Suppose SE is an object of SummarizedExperiment class
slotClass(SE)
slotClass(SE, "colData")
}
}
