% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2matrix.R
\name{table2matrix}
\alias{table2matrix}
\title{Convert a table object to a matrix}
\usage{
table2matrix(tbl)
}
\arguments{
\item{tbl}{a table object}
}
\description{
Convert a table object to a matrix
}
\examples{
\dontrun{

a = sample(letters[1:5], 20, replace = T) 
b = sample(letters[1:5], 20, replace = T) 
tbl = table(a, b)
View(tbl)
mat = table2matrix(tbl)
View(mat)
}
}
