% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_Publication}
\alias{theme_Publication}
\title{ggplot theme for publication ready Plots}
\usage{
theme_Publication(base_size = 14, base_family = c("Helvetica", "sans"),
  x_angle = 0, x_hjust = 0.5, x_vjust = 0.5, y_angle = 0,
  y_hjust = 0.5, y_vjust = 0.5, legend.key.size = unit(0.3, "cm"))
}
\arguments{
\item{base_size}{font base size.}

\item{base_family}{font base family.}

\item{x_angle}{x label angle.}

\item{x_hjust}{x horizontal alignment.}

\item{x_vjust}{x vertical alignment.}

\item{y_angle}{y label angle.}

\item{y_hjust}{y horizontal alignment.}

\item{y_vjust}{y vertical alignment.}

\item{legend.key.size}{legend key size.}
}
\description{
ggplot theme for publication ready Plots
}
