% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd2thisFile.R
\name{thisFile}
\alias{thisFile}
\title{The path, directory and file name of the current (executed) file}
\usage{
thisFile()
}
\value{
a list of full path, directory and file name of the current 
(executed) file.
}
\description{
The path, directory and file name of the current (executed) file
}
\examples{
{
\dontrun{
# Put the following codes in one file and save.
thisFile()
# Or run it in the R studio console.
}
}
}
