% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyText.R
\name{tidyText}
\alias{tidyText}
\title{Print text in a tidy way.}
\usage{
tidyText(
  text,
  leftSign = "\\\\{",
  commaSign = ",",
  rightSign = "\\\\}",
  showN = 1000,
  removeSpace = TRUE
)
}
\arguments{
\item{text}{the text to be tidy.}

\item{leftSign}{the left one of paired sign.}

\item{commaSign}{the unpaired sign, default is ",".}

\item{rightSign}{the right one of paired sign.}

\item{showN}{integer, the number of characters to show.}

\item{removeSpace}{logical, remove unnecessary space, the default is TRUE.}
}
\description{
Print text in a tidy way.
}
\examples{
text = 'list1:{A:{a = 1, b = 3, c = "4"}, B:{x = "", y = "1-3"}}'
tidyText(text)

}
