% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Truthy}
\alias{Truthy}
\title{Truthy}
\usage{
Truthy(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
\code{Truthy()}
}
\examples{
# Returns if a value exists or not:
Truthy(TRUE) # TRUE
Truthy(FALSE) # FALSE
Truthy(NULL) # FALSE
Truthy(NA) # FALSE
}

