% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Vapply.R
\name{Vapply}
\alias{Vapply}
\title{Wrapper for vapply function.}
\usage{
Vapply(.x, .f, fun_value, ..., use_names = TRUE)
}
\arguments{
\item{.x}{A vector.}

\item{.f}{A function to be applied.}

\item{fun_value}{A (generalized) vector; a template for the return value from .f.}

\item{...}{Optional arguments to .f.}

\item{use_names}{Logical; if TRUE and if X is character, use .x as names for the result unless it had names already.}
}
\description{
Wrapper for \code{\link[base]{vapply}}.
}
\seealso{
\code{\link[base]{vapply}} for code and documentation.

Other functionals: \code{\link{Apply}};
  \code{\link{Lapply}}; \code{\link{Mapply}};
  \code{\link{Negate}}; \code{\link{Sapply}};
  \code{\link{Tapply}}
}

