% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where2SearchFun.R
\name{whereis}
\alias{whereis}
\title{Which environment/package(s) is/are a function located in the search() path}
\usage{
whereis(x, env = parent.frame(), mode = c("function", "any"))
}
\arguments{
\item{x}{function or function name to search.}

\item{env}{environment to search, if the function x is not found in this 
environment, the parent environment is further searched until the top 
environment, i.e. empty environment (\code{emptyenv()}) 
or \code{R_EmptyEnv}.}

\item{mode}{either "function" or "any", if mode = "function" (default),
considering x is only a function object.}
}
\value{
environment where the function/object x is found.
If x is not found, \code{NULL} is returned. If multiple 
environments are found, the first environment is the default place to 
get the value of x.
}
\description{
Which environment/package(s) is/are a function located in the search() path
}
\examples{
\dontrun{
detachPackages("Matrix")
whereis("print")
whereis(print)

# Matrix::print is the default print
library(Matrix)
whereis("print")
identical(print, Matrix::print)
identical(print, base::print)

print = function() print()
whereis("print")

# Search any object
whereis(.Options, mode = "any")
}
}
