% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Withdraw}
\alias{Withdraw}
\title{Withdraw}
\usage{
Withdraw(obj, fields)
}
\arguments{
\item{obj}{a predicate function.}

\item{fields}{a character vector of fields.}
}
\value{
the negation of that function.
}
\description{
Withdraw
}
\examples{
# Create a function, compact(), that removes all null elements from a list:
new_model <- lm(mtcars, formula = hp ~ wt)
getCoefficients <- Plucker("coefficients")
getCoefficients(new_model)
}

