% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{All}
\alias{All}
\title{All}
\usage{
All(f, x, na.rm = FALSE)
}
\arguments{
\item{f}{a predicate function.}

\item{x}{an object.}
}
\value{
a logical value.
}
\description{
\code{All()}
}
\examples{
# comment here
All(mtcars, is.numeric) # TRUE
All(mtcars, is.character) # FALSE
mtcars$am <- factor(mtcars$am)
All(mtcars, is.numeric) # FALSE
All(mtcars, is.factor) # FALSE

# Handles NAs and NULLs
All(list(NA, "3", NULL), is.numeric) # FALSE
All(list(NA, "3", NULL, 5), is.numeric) # FALSE

# Use na.rm = TRUE to remove NULLS
All(list(NA, TRUE), Identity) # NA
All(list(NA, TRUE), Identity, na.rm = TRUE) # TRUE
}

