% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{All}
\alias{All}
\title{All}
\usage{
All(x, f)
}
\arguments{
\item{x}{an object.}

\item{f}{a predicate function.}
}
\value{
a logical value.
}
\description{
\code{All()}
}
\examples{
# comment here
All(mtcars, is.numeric) # TRUE
All(mtcars, is.character) # FALSE
mtcars$am <- factor(mtcars$am)
All(mtcars, is.numeric) # FALSE
All(mtcars, is.factor) # FALSE
}

