% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Always.R
\name{Always}
\alias{Always}
\title{Create a function that that always returns a specific object.}
\usage{
Always(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a function that itself returns x.
}
\description{
\code{Always(x)} is a closure function that always returns a function
returning object x.
}
\examples{
# comment here
always_0 <- Always(0)
always_0() # 0
always_true <- Always(TRUE)
always_true() # TRUE
}

