% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/And.R
\name{And}
\alias{And}
\title{And}
\usage{
And(f, g)
}
\arguments{
\item{f}{a predicate function.}

\item{g}{a predicate function.}
}
\value{
a function.
}
\description{
\code{And()}
}
\examples{
# comment here
}

