% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Andify.R
\name{Andify}
\alias{Andify}
\title{Andify}
\usage{
Andify(f, g)
}
\arguments{
\item{f}{a predicate function.}

\item{g}{a predicate function.}
}
\value{
a function.
}
\description{
\code{Andify()}
}
\examples{
# comment here
}

