% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Any}
\alias{Any}
\title{Any}
\usage{
Any(x, f)
}
\arguments{
\item{x}{an object.}

\item{f}{a predicate function.}
}
\value{
a logical value.
}
\description{
\code{Any()}
}
\examples{
# comment here
Any(mtcars, is.numeric) # TRUE
Any(mtcars, is.character) # FALSE
mtcars$am <- factor(mtcars$am)
Any(mtcars, is.numeric) # TRUE
Any(mtcars, is.factor) # TRUE
}

