% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Any}
\alias{Any}
\title{Any}
\usage{
Any(f, x, na.rm = FALSE)
}
\arguments{
\item{f}{a predicate function.}

\item{x}{an object.}
}
\value{
a logical value.
}
\description{
\code{Any()}
}
\examples{
# comment here
Any(is.numeric, mtcars) # TRUE
Any(is.character, mtcars) # FALSE
mtcars$am <- factor(mtcars$am)
Any(is.numeric, mtcars) # TRUE
Any(is.factor, mtcars) # TRUE

# Handles NAs and NULLs
Any(is.numeric, list(NA, "3", NULL)) # FALSE
Any(is.numeric, list(NA, "3", NULL, 5)) #TRUE

# Use na.rm = TRUE to remove NULLS
Any(Identity, list(NA, FALSE)) # NA
Any(Identity, list(NA, FALSE), na.rm = TRUE) # FALSE
}

