% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Apply.R
\name{Apply}
\alias{Apply}
\title{Apply}
\usage{
Apply(f, x, margin, ...)
}
\arguments{
\item{f}{the function to be applied.}

\item{x}{an array, including a matrix.}

\item{margin}{a vector giving the subscripts which the function will be applied over.}

\item{...}{optional arguments to f.}
}
\value{
something.
}
\description{
\code{Apply}
}
\examples{
# Some examples
}

