% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Apply.R
\name{Apply}
\alias{Apply}
\title{Wrapper for apply function.}
\usage{
Apply(.x, .f, .margin, ...)
}
\arguments{
\item{.x}{An array, including a matrix.}

\item{.f}{A function to be applied.}

\item{.margin}{A vector giving the subscripts which the function will be applied over.}

\item{...}{Optional arguments to f.}
}
\description{
Wrapper for \code{\link[base]{apply}}.
}
\seealso{
\code{\link[base]{apply}} for code and documentation.

Other functionals: \code{\link{Lapply}};
  \code{\link{Mapply}}; \code{\link{Negate}};
  \code{\link{Sapply}}; \code{\link{Tapply}};
  \code{\link{Vapply}}
}

