% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Best.R
\name{Best}
\alias{Best}
\title{Best}
\usage{
Best(f, x)
}
\arguments{
\item{x}{a vector.}
}
\value{
a two-argument function that returns a logical constant.
}
\description{
\code{Best()} takes a vector and a two-argument function that returns a logical constant (either TRUE or FALSE),
and returns the that function reduced over that vector.
}
\examples{
# Simulate the behavior of max with numerics
Best(1:10, function(x, y) return(x > y))
# Simulate the behavior of min with numerics
Best(1:10, function(x, y) return(x < y))
# This comparison function prefers values that begin with l
Best(letters, function(x, y) return(x[1] == "l"))
}

