% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Best.R
\name{Best}
\alias{Best}
\title{Best}
\usage{
Best(f, x)
}
\arguments{
\item{f}{a function.}

\item{x}{a value.}
}
\value{
a two-argument function that returns a logical constant.
}
\description{
\code{Best()} takes a vector and a two-argument function that returns
a logical constant (either TRUE or FALSE), and returns that function
reduced over that vector.
}
\examples{
# Simulate the behavior of max with numerics
Best(function(x, y) return(x > y), 1:10)
# Simulate the behavior of min with numerics
Best(function(x, y) return(x < y), 1:10)
# This comparison function prefers values that begin with l
Best(function(x, y) return(x[1] == "l"), letters)
}
\seealso{
Other aggregate functions: \code{\link{Finder}};
  \code{\link{Max}}; \code{\link{Min}}
}

