% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Best.R
\name{Best}
\alias{Best}
\title{Find the best value in a vector.}
\usage{
Best(f, x)
}
\arguments{
\item{f}{A binary predicate function.}

\item{x}{A vector.}
}
\value{
The best value in that vector, as determined by the binary predicate function.
}
\description{
\code{Best()} takes a vector \code{x} and a binary predicate function
\code{f} and returns the result of \code{f} reduced over \code{x}.
}
\examples{
# Simulate the behavior of max with numerics
Best(function(x, y) return(x > y), 1:10)
# Simulate the behavior of min with numerics
Best(function(x, y) return(x < y), 1:10)
# This comparison function prefers values that begin with l
Best(function(x, y) return(x[1] == "l"), letters)
}
\seealso{
Other aggregate functionals: \code{\link{Mapply}};
  \code{\link{Max}}; \code{\link{Min}};
  \code{\link{Sapply}}; \code{\link{Split}};
  \code{\link{Tapply}}; \code{\link{Vapply}}
}

