% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbindList.R
\name{cbindList}
\alias{cbindList}
\title{cbind all elements in a list}
\usage{
cbindList(List, ID = 0, replace = NA, sep = ".", fast = FALSE)
}
\arguments{
\item{List}{a list of data frame.}

\item{ID}{The number of column, or the column name that need to be refered.
If ID = 0, then the row names of each data frame will be used.}

\item{replace}{The value to replace the NAs that generated by cbind of data frame.}

\item{sep}{seperation character of the column names in the data frame
and the corresponding list index. The default is sep = ".".}

\item{fast}{using cbind if the rownames of each element in the \code{List}
are the same. The default is FALSE.}

\item{includeColId}{Logical. Include the refered column if it is TRUE}
}
\value{
A combined data frame, in which the colnames are the colnames
in the original data frame and corresponding list index
}
\description{
merge a list in which are all data.frame with one common
rowname or one common column number
}
\examples{
{
\dontrun{
# The zeros/ones matrix
a = data.frame(a = letters[1:5], b = rnorm(5))
b = data.frame(a = letters[3:7], c = rnorm(5))
x = list(a, b)
cbindList(List = x, ID = 1)
cbindList(List = x, ID = 0)
cbindList(List = x, ID = "a")
cbindList(List = x, ID = "a", replace = 0)
cbindList(List = x, ID = "a", replace = 0, sep = "_")
}
}
}
