% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckBiocReports.R
\name{CheckBiocReports}
\alias{CheckBiocReports}
\title{Check the package reports on Biocondocutor}
\usage{
CheckBiocReports(
  package = "RegEnrich",
  url0 = "http://bioconductor.org/checkResults/devel/bioc-LATEST/"
)
}
\arguments{
\item{package}{the package name}

\item{url0}{the url of the latest bioconductor package reports.}
}
\description{
Check the package reports on Biocondocutor
}
\examples{
\dontrun{
CheckBiocReports()
CheckBiocReports("FlowSOM")
}
}
