% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd2.R
\name{clear}
\alias{clear}
\title{clear}
\usage{
clear(pattern = "*", all.names = FALSE, envir = parent.frame())
}
\arguments{
\item{pattern}{A character string. The regular expression pattern of
variable names to be removed.}

\item{all.names}{Logical. If TRUE, all object names are returned.
If FALSE, names which begin with a . are omitted.}

\item{envir}{An alternative argument to name for specifying the environment.}
}
\value{
The name of variables which have been removed.
}
\description{
clear removes the variables, matched by the pattern,
from the environment (the default is Global Environment).
It will return the name of variables which have been removed.
}
\examples{
# Remove all (visible) variables in the Global Environment.
x1 = 1
x3 = "a"
xx = TRUE
y = function(x) x
y1 = 1:4
yx = list(y = 1, x = 3)
.x = data.frame(rnorm(5))
ls(all.names = TRUE)
clear()
ls(all.names = TRUE)

# Remove all variables including (invisible) variables begin with a ".".
clear(all.names = TRUE)
ls(all.names = TRUE)

# Remove variables begin with "x".
x1 = 1
x3 = "a"
xx = TRUE
y = function(x) x
y1 = 1:4
yx = list(y = 1, x = 3)
.x = data.frame(rnorm(5))
clear("^x")
ls(all.names = TRUE)

# Remove variables end with "x".
clear("x$")
ls(all.names = TRUE)

}
