% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Compact}
\alias{Compact}
\title{Compact}
\usage{
Compact(x)
}
\arguments{
\item{x}{a vector.}
}
\description{
\code{Compact()} takes a vector x and returns it with all NULL values filtered out.
}
\examples{
# Removes all null elements from a vector:
a <- list(NULL, 1, 5, NULL)
Compact(a)

b <- c(1, 2, 0, 4, NULL, 1, 3, NULL)
Compact(b)
}

