% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Compose}
\alias{Compose}
\title{Compose}
\usage{
Compose(f, g)
}
\arguments{
\item{f}{a predicate function.}
}
\value{
the negation of that function.
}
\description{
\code{Failwith()} turns a function that throws an error into a function that returns a default value when there’s an error.
The essence of failwith() is simple; it’s just a wrapper around try(), the function that captures errors and allows execution to continue.
}
\examples{
# Compose length and unique
lu <- Compose(length, unique)
lu(c(1:10, 5:15, 20:25))
}

