% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Compose.R
\name{Compose}
\alias{Compose}
\alias{\%O\%}
\title{Compose multiple functions.}
\usage{
Compose(...)

f \%O\% g
}
\arguments{
\item{...}{n functions to apply in order from right to left.}

\item{f,g}{Two functions to compose for the infix form}
}
\value{
A function that
}
\description{
In infix and prefix forms.
}
\examples{
not_null <- `!` \%O\% is.null
not_null(4)
not_null(NULL)

add1 <- function(x) x + 1
Compose(add1,add1)(8)
}
\seealso{
Other function operators: \code{\link{Fail_With}}
}

