% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Failwith.R
\name{Failwith}
\alias{Failwith}
\title{Failwith}
\usage{
Failwith(default = NULL, f, silent = FALSE)
}
\arguments{
\item{default}{default value.}

\item{f}{any function that throws an error.}

\item{silent}{logical: should the report of error messages be suppressed?.}
}
\value{
a function that returns a default value when there's an error.
}
\description{
Failwith() turns a function that throws an error into a function that
returns a default value when there is an error. The essence of Failwith()
is simple: it is just a wrapper around try(), the function that
captures errors and allows execution to continue.
}
\examples{
# Some examples
}
\seealso{
Other function operators: \code{\link{Compose}},
  \code{\link{\%O\%}}
}

