% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Constants.R
\name{False}
\alias{False}
\title{Return the value FALSE.}
\usage{
False()
}
\value{
FALSE.
}
\description{
\code{False()} is a function that returns FALSE.
}
\examples{
# False() returns FALSE:
False()
}
\seealso{
Other constants: \code{\link{Identity}}; \code{\link{Na}};
  \code{\link{Null}}; \code{\link{True}}
}

