% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq2datatable.R
\name{fastq2dataframe_rev}
\alias{fastq2dataframe_rev}
\title{reverse a fastq corresponding data frame to a data.table of fastq format, and save a fastq file}
\usage{
fastq2dataframe_rev(fastqDataframe, fileName = NULL)
}
\arguments{
\item{fastqDataframe}{a data frame of four columns, and the column names must be
"label", "sequence", "mark", and "score".}

\item{fileName}{the fastq file to be written.}
}
\description{
reverse a fastq corresponding data frame to a data.table of fastq format, and save a fastq file
}
\examples{
{
\dontrun{
# Generating information for 3 reads, of which is length of 50.
len = 50
n = 3
sequences = sapply(1:n, function(x) paste0(sample(c("A", "T", "C", "G"), len, replace = T), collapse = ""))
base = 33:(33+42) # ASCII_BASE=33
mode(base) = "raw"
scores = sapply(1:n, function(x) paste0(sapply(sample(base[2:40], len, replace = T), rawToChar), collapse = ""))
fastq_test = data.frame(label = paste0("sequence", 1:n),
                        sequence = sequences,
                        mark = "+", score = scores)
# save file
fastq2dataframe_rev(fastq_test, "./fastq_test.fastq")
}
}
}
\seealso{
\link{\code{fastq2dataframe}}
}
