% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionInfo.R
\name{funCode}
\alias{funCode}
\alias{funCode.default}
\alias{funCode.ggproto_method}
\title{Finding function source code\cr}
\usage{
funCode(f, ...)

\method{funCode}{default}(f = character(), pattern = NULL, envir = topenv(parent.frame()))

\method{funCode}{ggproto_method}(f)
}
\arguments{
\item{f}{a function name with or without quotation}

\item{pattern}{regular expression to match the object class.}

\item{envir}{environment to search}
}
\value{
A \code{\link{functionInfo}} object, containing the function name,
function type, whether containing S3 or S4 methods, dirrect function source,
S3 methods info, and S4 methods info.
}
\description{
funCode function is to obtain the source code of an R function
}
\examples{
\dontrun{
# The source of print
funCode(print)
funCode("print")

# The source of funCode
funCode(funCode)
funCode(lm)

# The source of plot for data.frame
funCode("plot", "data.frame")

# The source of standardGeneric show
funCode("show")
funCode(show)

# The source of ggproto_method in ggplot2
library(ggplot2)
funCode(Layout$setup)
}
}
