% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionInfo.R
\docType{class}
\name{functionInfo-class}
\alias{functionInfo-class}
\alias{functionInfo}
\alias{show,functionInfo-method}
\title{Generating functionInfo object}
\usage{
\S4method{show}{functionInfo}(object)
}
\arguments{
\item{object}{an object of functionInfo class}

\item{funName}{character, the function name.}

\item{typeof}{character, the result generated by `typeof`` function.}

\item{S3S4}{logical, whether the results of \code{\link{.S3methods}}
and \code{\link{.S4methods}} exist.}

\item{fcn}{function, the result shown by `print.function` directly.}

\item{fS3}{fcnS object, the information of S3 methods.}

\item{fS4}{fcnS object, the information of S4 mehtods.}
}
\value{
a functionInfo object
}
\description{
FunctionInfo is an S4 class and a function to generate functionInfo object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show(functionInfo)}: Show the S3 method names in
object@fS3 and S4 method names in object@fS4

}}
\examples{
\dontrun{
functionInfo()
}
}
