% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\docType{package}
\name{functools}
\alias{functools}
\alias{functools-package}
\title{functools: An R package for extending functional programming in R.}
\description{
functools is a utility-belt library for R that adds support to the
usual higher order functional suspects (Map, Reduce, Filter, etc.) without
extending any core R objects.
}
\details{
It has three main types of helper functions:

\itemize{
\item Closures
\item Functionals
\item Function Operators
}

To learn more about functools, start with the vignettes:
\code{browseVignettes(package = "functools")}
}

