% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\docType{functools}
\name{functools}
\alias{functools}
\title{functools: A package for implementing Higher Order Functions in R.}
\description{
functools: A package for implementing Higher Order Functions in R.
}

