% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionInfo.R
\name{funDispatch}
\alias{funDispatch}
\title{find dispatched functions for an object}
\usage{
funDispatch(f, object, envir = topenv(parent.frame()))
}
\arguments{
\item{f}{function name.}

\item{object}{the object that the function f is processing}

\item{envir}{the working environment}
}
\description{
find dispatched functions for an object
}
