% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lapply.R
\name{Lapply}
\alias{Lapply}
\title{Lapply}
\usage{
Lapply(f, x, ...)
}
\arguments{
\item{f}{a function.}

\item{check}{a function.}

\item{init}{a value.}
}
\value{
a list.
}
\description{
\code{Lapply}
}
\examples{
Some examples:
}

