% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lapply.R
\name{Lapply}
\alias{Lapply}
\title{Wrapper for lapply function.}
\usage{
Lapply(.x, .f, ...)
}
\arguments{
\item{.x}{A vector.}

\item{.f}{A function to be applied.}

\item{...}{Optional arguments to .f.}
}
\description{
Wrapper for \code{\link[base]{lapply}}.
}
\seealso{
\code{\link[base]{lapply}} for code and documentation.

Other functionals: \code{\link{Apply}};
  \code{\link{Mapply}}; \code{\link{Map}};
  \code{\link{Sapply}}; \code{\link{Tapply}};
  \code{\link{Vapply}}
}

