% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{make_call}
\alias{make_call}
\title{make_call}
\usage{
make_call(f, ..., .args = list())
}
\arguments{
\item{x}{a vector.}
}
\description{
\code{Compact()} takes a vector x and returns it with all NULL and NA values filtered out.
}
\examples{
Some examples:
}

