% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Mapply.R
\name{Mapply}
\alias{Mapply}
\title{Mapply}
\usage{
Mapply(f, x)
}
\arguments{
\item{f}{a predicate function.}

\item{x}{a value.}
}
\value{
something.
}
\description{
\code{Mapply()}
}
\examples{
# Some examples
}
\seealso{
Other aggregate functionals: \code{\link{Best}};
  \code{\link{Max}}; \code{\link{Min}};
  \code{\link{Sapply}}; \code{\link{Split}};
  \code{\link{Tapply}}; \code{\link{Vapply}}
}

