% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Max.R
\name{Max}
\alias{Max}
\title{Max}
\usage{
Max(f, x)
}
\arguments{
\item{f}{a predicate function.}

\item{x}{a value.}
}
\value{
something.
}
\description{
\code{Max()}
}
\examples{
# Some examples
}
\seealso{
Other aggregate functions: \code{\link{Finder}};
  \code{\link{Min}}
}

