% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Constants.R
\name{Na}
\alias{Na}
\title{Return the value NA.}
\usage{
Na()
}
\value{
NA.
}
\description{
\code{Na()} is a function that returns NA.
}
\examples{
# Na() returns NA:
Na()
}
\seealso{
Other constants: \code{\link{False}};
  \code{\link{Identity}}; \code{\link{Null}};
  \code{\link{True}}
}

