% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Negate}
\alias{Negate}
\title{negate}
\usage{
Negate(f)
}
\arguments{
\item{f}{a predicate function.}
}
\value{
the negation of that function.
}
\description{
\code{negate()} takes a function that returns a logical vector (a predicate function), and returns the negation of that function.
This can be a useful shortcut when a function returns the opposite of what you need.
}
\examples{
# Create a function, compact(), that removes all null elements from a list:
compact <- function(x) Filter(Negate(is.null), x)
foo <- list(NULL, 1, 5, NULL)
compact(foo)
}

