% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Negate}
\alias{Negate}
\title{Negate}
\usage{
Negate(f)
}
\arguments{
\item{f}{a predicate function.}
}
\value{
the negation of that function.
}
\description{
\code{negate()} takes a function that returns a logical vector (a predicate function), and returns the negation of that function.
This can be a useful shortcut when a function returns the opposite of what you need.
}
\examples{
# Create a function, compact(), that removes all null elements from a list:
compact <- function(x) Filter(Negate(is.null), x)
foo <- list(NULL, 1, 5, NULL)
compact(foo)

# Create a function that evalues the even-ness of a number
is_even <- function(n) return(n \%\% 2 == 0)
is_even(4)
is_even(5)
# Use Negate to create a function that evaluates the odd-ness of a number
is_odd <- Negate(is_even)
is_odd(4)
is_odd(5)
}

