% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Orify.R
\name{Orify}
\alias{Orify}
\title{Create new predicate functions}
\usage{
Orify(...)
}
\arguments{
\item{...}{n functions to apply in order from right to left}
}
\description{
Create new predicate functions
}
\examples{
# Examples
}

