% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functools.R
\name{Partial}
\alias{Partial}
\title{Partial apply a function, filling in some arguments.}
\usage{
Partial(`_f`, ..., .env = parent.frame(), .lazy = TRUE)
}
\arguments{
\item{_f}{a function. For the output source to read well, this should be an be a named function. This argument has the weird (non-syntactic) name _f so it doesn't accidentally capture any argument names begining with f.}

\item{...}{named arguments to f that should be partially applied.}

\item{.env}{the environment of the created function. Defaults to parent.frame and you should rarely need to modify this.}

\item{.lazy}{If TRUE arguments evaluated lazily, if FALSE, evaluated when partial is called.}
}
\description{
Partial function application allows you to modify a function by pre-filling some of the arguments. It is particularly useful in conjunction with functionals and other function operators.
}
\examples{
Some examples:
}

