% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plucker.R
\name{Plucker}
\alias{Plucker}
\title{Plucker}
\usage{
Plucker(field)
}
\arguments{
\item{f}{a predicate function.}
}
\value{
the negation of that function.
}
\description{
\code{negate()} takes a function that returns a logical vector (a predicate function), and returns the negation of that function.
This can be a useful shortcut when a function returns the opposite of what you need.
}
\examples{
# Pluck a value from a list or S3 object:
new_model <- lm(mtcars, formula = hp ~ wt)
getCoefficients <- Plucker("coefficients")
getCoefficients(new_model)
}

