% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plucker.R
\name{Plucker}
\alias{Plucker}
\title{Plucker}
\usage{
Plucker(field)
}
\arguments{
\item{field}{a character vector.}
}
\value{
something.
}
\description{
Pluck objects from S3 and S4 object slots!
}
\examples{
# Pluck a value from a list or S3 object:
new_model <- lm(mtcars, formula = hp ~ wt)
getCoefficients <- Plucker("coefficients")
getCoefficients(new_model)
}

