% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub.R
\name{qdelAll}
\alias{qdelAll}
\title{qdel of job by pattern}
\usage{
qdelAll(pattern = "*")
}
\arguments{
\item{pattern}{The pattern of job IDs.}
}
\value{
Deleted job IDs.
}
\description{
qdel of job by pattern
}
\seealso{
\code{\link{qdel}}
}
