% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub_slurm.R
\name{qstat2}
\alias{qstat2}
\title{Status (sacct) of SLURM. jobs.}
\usage{
qstat2(
  stat = c("running", "completed", "failed", "timeout", "resizing", "deadline",
    "node_fail"),
  groups = NULL,
  users = NULL
)
}
\arguments{
\item{stat}{the job status, including "running", "completed", "failed", 
"timeout", "resizing", "deadline", "node_fail".}
}
\description{
Status (sacct) of SLURM. jobs.
}
\seealso{
\code{\link{qstat}}
}
