% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub.R
\name{qstatAll}
\alias{qstatAll}
\title{Status (qstat) of running jobs of all users.}
\usage{
qstatAll(stat = c("all", "run", "wait"))
}
\arguments{
\item{stat}{the job status, including "all", "run", and "wait".}
}
\value{
a data frame of job details, including job ID, prior, name, user,
submit.start, at, queue, slots, and ja.task.ID.
}
\description{
Status (qstat) of running jobs of all users.
}
\seealso{
\code{\link{qstat}}, \code{\link{qstatGroupAll}}
}
