% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_switch.R
\name{R_switch}
\alias{R_switch}
\title{Switch R framework version on Mac}
\usage{
R_switch(
  version = R_frameworks(),
  reset = TRUE,
  openRStudio = TRUE,
  frameworkPath = "/Library/Frameworks/R.framework/",
  RStudioPath = "/Applications/RStudio.app/Contents/MacOS/RStudio"
)
}
\arguments{
\item{reset}{if TRUE, the framework is changed back to the
original framework version.}

\item{openRStudio}{if TRUE, open an RStudio session with the new R framework
version.}

\item{frameworkPath}{the R.framework path, the default is 
"/Library/Frameworks/R.framework/" on Mac.}

\item{RStudioPath}{RStudio path, the default is 
"/Applications/RStudio.app/Contents/MacOS/RStudio".}

\item{Version}{R framework version to switch to}
}
\description{
Switch R framework version on Mac
}
\examples{
\dontrun{
# List all frameworks
fw = R_frameworks()
fw

# Current framework
cfw = R_currentFramework()
cfw

# Switch
R_switch(setdiff(fw, cfw)[1], reset = FALSE, openRStudio = TRUE)
R_switch(cfw, reset = FALSE, openRStudio = TRUE)

R_switch(setdiff(fw, cfw)[1], reset = TRUE, openRStudio = FALSE)

# Reset the initial framework
R_switch(cfw, reset = FALSE)
}
}
