% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Reduce-Right.R
\name{Reduce_Right}
\alias{Reduce_Right}
\title{Simple wrapper for Reduce, proceeding from the right.}
\usage{
Reduce_Right(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[base]{Reduce}}.}
}
\description{
Wrapper for \code{\link[base]{Reduce}} with \code{right} set to \code{TRUE}.
}
\seealso{
\code{\link[base]{Reduce}} for code and documentation.

Other function operators: \code{\link{Compose}},
  \code{\link{\%O\%}}; \code{\link{Fail_With}};
  \code{\link{Memoise}}; \code{\link{Partial}}
}

