% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Repeatedly.R
\name{Repeatedly}
\alias{Repeatedly}
\title{Repeatedly}
\usage{
Repeatedly(f, x)
}
\arguments{
\item{f}{a predicate function.}

\item{x}{a value.}
}
\value{
something.
}
\description{
Explanation
}
\examples{
# Some examples
}

