% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub.R
\name{rQsub}
\alias{rQsub}
\title{Submit (qsub) R jobs.}
\usage{
rQsub(path = getwd(), rFile = "testQsub.R", jobName = "job",
  threaded = 1, memoryG = 10, rTimeHour = 2,
  logFile = "logfilename.log", email = NULL, when2Email = "aes",
  preCMD = "echo \\"module load R/3.3.0 && Rscript ", param1 = 1:10,
  ...)
}
\arguments{
\item{path}{file path.}

\item{rFile}{the R script to run.}

\item{jobName}{the job name(s).}

\item{threaded}{integer or integer vector to specify
the number of threads of each job.}

\item{memoryG}{numeric or numeric vector to specify the
memory (in Gb) of each job.}

\item{rTimeHour}{numeric or numeric vector. Maximum running
time (in hour) of each job.}

\item{logFile}{the log file(s) for errors and warnings.}

\item{email}{email to receive report from HPC. The default is NULL.}

\item{when2Email}{when will the server send a email. This can be
"b" (start) and/or "e" (end) and/or "a" (abortion) and/or "s" (suspension)
of your job. The default is "aes". This parameter will not be
used if email is NULL.}

\item{preCMD}{the command to run R script, the default is
'echo "module load R/3.3.0 && Rscript '.}

\item{param1}{the first parameter passed to R script.}

\item{...}{other parameter passed to R script.}
}
\value{
A list of output of running rFile.
}
\description{
Submit R jobs in parallel (array).
}
