% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQsub.R
\name{rQsub}
\alias{rQsub}
\title{Submit (qsub) R jobs.}
\usage{
rQsub(path = getwd(), rFile = paste0(path, "/testQsub.R"),
  jobName = "job", threaded = 1, memoryG = 10, rTimeHour = 2,
  logFile = paste0(path, "/logfilename.log"),
  email = "wtao@umcutrecht.nl",
  preCMD = "echo \\"module load R/3.3.0 && Rscript ", param1 = 1:10,
  ...)
}
\arguments{
\item{path}{file path.}

\item{rFile}{the R script to run.}

\item{jobName}{the job name(s).}

\item{threaded}{the number of threads.}

\item{memoryG}{memory in Gb.}

\item{rTimeHour}{maximum running time in hour.}

\item{logFile}{the log file(s) for errors and warnings.}

\item{email}{email to receive report from HPC.}

\item{preCMD}{the command to run R script, the default is
'echo "module load R/3.3.0 && Rscript '.}

\item{param1}{the first parameter passed to R script.}

\item{...}{other parameter passed to R script.}
}
\value{
A list of output of running rFile.
}
\description{
Submit R jobs in parallel (array).
}
\examples{
{
\dontrun{
# This function can only run on HPC
# One parameter in R script
rQsub(path = "/my/file/path",
      rFile = paste0(path, "/testQsub.R"),
      jobName = "job", threaded = 1,
      memoryG = 10, rTimeHour = 2,
      logFile = paste0(path, "/logfilename.log"),
      preCMD = 'echo "module load R/3.3.0 && Rscript ',
      param1 = 1:10)
# Multiple parameters in R script and using 2 threads
rQsub(path = "/my/file/path",
      rFile = paste0(path, "/testQsub.R"),
      jobName = "job", threaded = 2,
      param1 = 1:10, param2 = 2:11, param3 = 3:12)
}
}
}
