% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Sapply.R
\name{Sapply}
\alias{Sapply}
\title{Wrapper for sapply function.}
\usage{
Sapply(.x, .f, ..., simplify = TRUE, use_names = TRUE)
}
\arguments{
\item{.x}{A vector.}

\item{.f}{A function to be applied.}

\item{...}{Optional arguments to f.}

\item{simplify}{Logical or character string; should the result be simplified to a vector, matrix or higher dimensional array if possible?}

\item{use_names}{Logical; if TRUE and if .x is character, use .x as names for the result unless it had names already.}
}
\description{
Wrapper for \code{\link[base]{sapply}}.
}
\seealso{
\code{\link[base]{sapply}} for code and documentation.

Other functionals: \code{\link{Apply}};
  \code{\link{Lapply}}; \code{\link{Mapply}};
  \code{\link{Negate}}; \code{\link{Tapply}};
  \code{\link{Vapply}}
}

