% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tapply.R
\name{Tapply}
\alias{Tapply}
\title{Tapply}
\usage{
Tapply(f, x)
}
\arguments{
\item{f}{a predicate function.}

\item{x}{a value.}
}
\value{
something.
}
\description{
\code{Tapply()}
}
\examples{
# Some examples
}
\seealso{
Other aggregate functionals: \code{\link{Best}};
  \code{\link{Max}}; \code{\link{Min}};
  \code{\link{Sapply}}; \code{\link{Vapply}}
}

