% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Constants.R
\name{True}
\alias{True}
\title{True}
\usage{
True()
}
\value{
TRUE.
}
\description{
\code{True()} is a function that returns TRUE.
}
\examples{
# True() returns TRUE:
True()
}
\seealso{
\code{link{True}}, \code{link{False}}, \code{link{Identity}}, \code{link{Null}}, \code{link{Na}},

Other constants: \code{\link{False}};
  \code{\link{Identity}}; \code{\link{Na}};
  \code{\link{Null}}
}

