% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Truthy.R
\name{Truthy}
\alias{Truthy}
\title{Truthy}
\usage{
Truthy(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
\code{Truthy()} returns TRUE or FALSE if an object is TRUE or not.
An object is is "TRUE" if it is exists and is TRUE.
}
\examples{
# Returns if a value exists or not:
Truthy(TRUE) # TRUE
Truthy(FALSE) # FALSE
Truthy(NULL) # FALSE
Truthy(NA) # FALSE
Truthy(2L) # TRUE
Truthy(1L) # TRUE
Truthy(0L) # FALSE
Truthy("a") # TRUE
}

