% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Vapply.R
\name{Vapply}
\alias{Vapply}
\title{Vapply}
\usage{
Vapply(f, x)
}
\arguments{
\item{f}{a predicate function.}

\item{x}{a value.}
}
\value{
something.
}
\description{
\code{Vapply()}
}
\examples{
# Some examples
}
\seealso{
Other aggregate functionals: \code{\link{Best}};
  \code{\link{Max}}; \code{\link{Min}};
  \code{\link{Sapply}}; \code{\link{Tapply}}
}

