% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Withdraw.R
\name{Withdraw}
\alias{Withdraw}
\title{Withdraw}
\usage{
Withdraw(obj, fields)
}
\arguments{
\item{obj}{a predicate function.}

\item{fields}{a character vector of fields.}
}
\value{
the negation of that function.
}
\description{
Withdraw
}
\examples{
Some examples:
}

